Bugfix in method compareTo() of PrimitiveValue.
This subtle bug caused two PrimitiveValues of numerical type to
incorrectly return the value 0 when their difference, in absolute
value, was less than 1.

This, in turn, would cause method equals() to incorrectly declare
that these two primitive values are the same object.

This, in turn, would cause the display of a table containing two
such values to merge cells that should not be merged, typically
causing lines of the tables to disappear from the display.